fix: move type shims to `next-env.d.ts` + fixed some type decs
